Shorten Input component
