+++ client/app/widgets/QuickCreate/index.jsx
@@ -2,7 +2,7 @@
 import React from 'react';
 import { Modal } from '../../components/Modal';
 import { Input } from '../../components/Input';
-import type { Checkbox } from '../../components/Input';
+import type { Checkbox } from '../../components/Input/utils';
 import { Utils } from '../../utils';
 import css from './QuickCreate.scss';
 import { DynamicForm } from '../../components/Form/DynamicForm';
